Update mongo_actions.sh, mongo_backup.tpl, and mongo_restore.tpl

+++ files/mongo_actions.sh
@@ -92,20 +92,6 @@ else
     exit 1
 fi
 
-### GET SOURCE DB CONNECTION DETAILS FROM SSM ###
-if [[ "${ACTION_TYPE}" == "mongo_clone" ]]; then
-    SDBNAME=$(aws ssm get-parameter --name "/infra/$INIT_DB_WORKSPACE/db-name" --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
-    SDBHOST=$(aws ssm get-parameter --name "/infra/$INIT_DB_WORKSPACE/db-host" --with-decryption --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
-    SDBUSER=$(aws ssm get-parameter --name "/infra/$INIT_DB_WORKSPACE/db-username" --with-decryption --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
-    SDBPASSWORD=$(aws ssm get-parameter --name "/infra/$INIT_DB_WORKSPACE/db-password" --with-decryption --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
-    
-    if [[ -z "$SDBUSER" ]] || [[ -z "$SDBPASSWORD" ]] || [[ -z "$SDBHOST" ]]; then
-        echo "Could not retrieve one or more parameters from SSM!!!"
-        exit 1
-    fi
-    
-fi
-
 ### GET TARGET DB CONNECTION DETAILS FROM SSM ###
 DBNAME=$(aws ssm get-parameter --name "/infra/$WORKSPACE/db-name" --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
 DBUSER=$(aws ssm get-parameter --name "/infra/$WORKSPACE/db-username" --with-decryption --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
@@ -159,6 +145,15 @@ mongo_backup() {
 
 mongo_clone() {
       echo "Copying init db..."
+      ### GET SSM PARAMS OF SOURCE DB ###
+      SDBNAME=$(aws ssm get-parameter --name "/infra/$INIT_DB_WORKSPACE/db-name" --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
+      SDBHOST=$(aws ssm get-parameter --name "/infra/$INIT_DB_WORKSPACE/db-host" --with-decryption --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
+      SDBUSER=$(aws ssm get-parameter --name "/infra/$INIT_DB_WORKSPACE/db-username" --with-decryption --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
+      SDBPASSWORD=$(aws ssm get-parameter --name "/infra/$INIT_DB_WORKSPACE/db-password" --with-decryption --query 'Parameter.Value' --profile $AWS_PROFILE  --output text)
+      if [[ -z "$SDBUSER" ]] || [[ -z "$SDBPASSWORD" ]] || [[ -z "$SDBHOST" ]]; then
+          echo "Could not retrieve one or more parameters from SSM!!!"
+          exit 1
+      fi
       [ ! "$(docker ps | grep mongodocker)" ] && docker run --name mongodocker -i -d mongo bash
       docker exec -i mongodocker mongodump --uri "$SDBHOST/$SDBNAME" -u$SDBUSER -p$SDBPASSWORD --gzip --archive | mongorestore --uri "$DBHOST" -u$DBUSER -p$DBPASSWORD --nsFrom="$SDBNAME.*" --nsTo="$DBNAME.*" --gzip --archive
       docker rm -f mongodocker
